Renamed some classes to make more sense. Added an entities class and changed Artefact, Character and Furniture to inherit from it. Removed duplicate 'Artefact' class and merged duplicate 'Location' classes.
