+++ .idea/libraries/Maven__com_alexmerz_graphviz_dot_parser_1_0.xml
@@ -1,0 +1,14 @@
+<component name="libraryTable">
+  <library name="Maven: com.alexmerz.graphviz:dot-parser:1.0" type="java-imported" external-system-id="Maven">
+    <properties groupId="com.alexmerz.graphviz" artifactId="dot-parser" version="1.0" />
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/com/alexmerz/graphviz/dot-parser/1.0/dot-parser-1.0.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/com/alexmerz/graphviz/dot-parser/1.0/dot-parser-1.0-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/com/alexmerz/graphviz/dot-parser/1.0/dot-parser-1.0-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
