+++ .idea/libraries/Maven__org_junit_platform_junit_platform_suite_api_1_8_2.xml
@@ -1,0 +1,14 @@
+<component name="libraryTable">
+  <library name="Maven: org.junit.platform:junit-platform-suite-api:1.8.2" type="java-imported" external-system-id="Maven">
+    <properties groupId="org.junit.platform" artifactId="junit-platform-suite-api" version="1.8.2" />
+    <CLASSES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/junit/platform/junit-platform-suite-api/1.8.2/junit-platform-suite-api-1.8.2.jar!/" />
+    </CLASSES>
+    <JAVADOC>
+      <root url="jar://$MAVEN_REPOSITORY$/org/junit/platform/junit-platform-suite-api/1.8.2/junit-platform-suite-api-1.8.2-javadoc.jar!/" />
+    </JAVADOC>
+    <SOURCES>
+      <root url="jar://$MAVEN_REPOSITORY$/org/junit/platform/junit-platform-suite-api/1.8.2/junit-platform-suite-api-1.8.2-sources.jar!/" />
+    </SOURCES>
+  </library>
+</component>
